refactor: 커스텀 예외의 디렉토리 구조 설정
