+++ src/main/java/wad/seoul_nolgoat/exception/auth/UnsupportedProviderException.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.exception;
+package wad.seoul_nolgoat.exception.auth;
 
 public class UnsupportedProviderException extends RuntimeException {
