+++ src/main/java/wad/seoul_nolgoat/exception/mapapi/TMapApiException.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.exception;
+package wad.seoul_nolgoat.exception.mapapi;
 
 public class TMapApiException extends RuntimeException {
