+++ src/main/java/wad/seoul_nolgoat/service/bookmark/BookmarkService.java
@@ -9,8 +9,8 @@ import wad.seoul_nolgoat.domain.store.StoreRepository;
 import wad.seoul_nolgoat.domain.user.User;
 import wad.seoul_nolgoat.domain.user.UserRepository;
 import wad.seoul_nolgoat.exception.ErrorMessages;
-import wad.seoul_nolgoat.exception.StoreNotFoundException;
-import wad.seoul_nolgoat.exception.UserNotFoundException;
+import wad.seoul_nolgoat.exception.notfound.StoreNotFoundException;
+import wad.seoul_nolgoat.exception.notfound.UserNotFoundException;
 
 @RequiredArgsConstructor
 @Service
