+++ src/main/java/wad/seoul_nolgoat/service/kakaoMap/KakaoMapService.java
@@ -11,8 +11,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-import wad.seoul_nolgoat.exception.AddressConversionException;
 import wad.seoul_nolgoat.exception.ErrorMessages;
+import wad.seoul_nolgoat.exception.mapapi.AddressConversionException;
 import wad.seoul_nolgoat.service.kakaoMap.dto.StoreAdditionalInfoDto;
 import wad.seoul_nolgoat.web.search.dto.CoordinateDto;
