+++ src/main/java/wad/seoul_nolgoat/service/review/ReviewService.java
@@ -10,9 +10,9 @@ import wad.seoul_nolgoat.domain.store.StoreRepository;
 import wad.seoul_nolgoat.domain.user.User;
 import wad.seoul_nolgoat.domain.user.UserRepository;
 import wad.seoul_nolgoat.exception.ErrorMessages;
-import wad.seoul_nolgoat.exception.ReviewNotFoundException;
-import wad.seoul_nolgoat.exception.StoreNotFoundException;
-import wad.seoul_nolgoat.exception.UserNotFoundException;
+import wad.seoul_nolgoat.exception.notfound.ReviewNotFoundException;
+import wad.seoul_nolgoat.exception.notfound.StoreNotFoundException;
+import wad.seoul_nolgoat.exception.notfound.UserNotFoundException;
 import wad.seoul_nolgoat.util.mapper.ReviewMapper;
 import wad.seoul_nolgoat.web.review.dto.request.ReviewSaveDto;
 import wad.seoul_nolgoat.web.review.dto.request.ReviewUpdateDto;
